Use a container rather than an image
The idea here is somewhat opinionated.

 1. start a container from an image and leave it running bash as a
    daemon
 2. grab the container id and store it in .rdo.docker.container_id
 3. use the container id with docker exec to run commands in the
    container persisting them to the running image
 4. commit to the same image after each command

So, this should replace the current image. I'm not sure if you hit the
limits of the docker revisions (128?) or not. I bet re-using the name is
a bad idea if someone just uses "ubuntu" for example.

Bad unintended consequences aside, using the container vs. an image
seems much better and gets you something closer to how you use
vagrant. The advantage here is that you can dev somewhat normally
whereas if you use images, you end can up with a pretty ridiculus
Dockerfile or script that gets run when the docker image gets built.